want class to support configuration
